Phive: upgrade to use latest version of Phive
The version previously used was not compatible with Xdebug 3.x, nor with PHP 8.1.

The current latest release - 0.15.0 - has a minimum PHP requirement of PHP 7.3, which matches this project's requirements, so should be safe to use.

If at some point in the future, the minimum PHP requirements would no longer match, the Phive version can be fixed to a specific release again.

Ref: https://github.com/phar-io/phive/releases